Move concerns to edition directory
As these concerns will only ever be used in conjunction with an
Edition, it makes sense to keep these namespaced under
`ContentBlock::Edition`
