+++ lib/engines/content_object_store/app/models/content_object_store/content_block/edition.rb
@@ -1,12 +1,12 @@
 module ContentObjectStore
   module ContentBlock
     class Edition < ApplicationRecord
-      include ContentObjectStore::Documentable
-      include ContentObjectStore::ValidatesDetails
-      include ContentObjectStore::HasAuthors
-      include ContentObjectStore::HasAuditTrail
-      include ContentObjectStore::Workflow
-      include ContentObjectStore::HasLeadOrganisation
+      include Documentable
+      include HasAuditTrail
+      include HasAuthors
+      include HasLeadOrganisation
+      include ValidatesDetails
+      include Workflow
     end
   end
 end
