+++ lib/engines/content_object_store/app/models/content_object_store/content_block/edition/documentable.rb
@@ -1,5 +1,5 @@
 module ContentObjectStore
-  module Documentable
+  module ContentBlock::Edition::Documentable
     extend ActiveSupport::Concern
 
     included do
