+++ lib/engines/content_object_store/app/models/content_object_store/content_block/edition/has_audit_trail.rb
@@ -1,5 +1,5 @@
 module ContentObjectStore
-  module HasAuditTrail
+  module ContentBlock::Edition::HasAuditTrail
     extend ActiveSupport::Concern
 
     def self.acting_as(actor)
