+++ lib/engines/content_object_store/app/models/content_object_store/content_block/edition/has_creator.rb
@@ -1,5 +1,5 @@
 module ContentObjectStore
-  module HasCreator
+  module ContentBlock::Edition::HasCreator
     extend ActiveSupport::Concern
 
     included do
