+++ lib/engines/content_object_store/app/models/content_object_store/content_block/edition/has_lead_organisation.rb
@@ -1,5 +1,5 @@
 module ContentObjectStore
-  module HasLeadOrganisation
+  module ContentBlock::Edition::HasLeadOrganisation
     extend ActiveSupport::Concern
 
     included do
