+++ lib/engines/content_object_store/app/models/content_object_store/content_block/edition/validates_details.rb
@@ -1,5 +1,5 @@
 module ContentObjectStore
-  module ValidatesDetails
+  module ContentBlock::Edition::ValidatesDetails
     extend ActiveSupport::Concern
 
     DETAILS_PREFIX = "details_".freeze
