+++ lib/engines/content_object_store/test/unit/app/models/content_block_edition/validates_details_test.rb
